Bug fixes
Added check for command being convertable to an exe to make_argv before
returing a malloced pointer either to a new full executable path, or
returning the same pointer to the first word in the token chain. This
was causing a segfault as the first word pointer was double freed.
Also fixed issue in export exporting the first argument which was export
itself with a value of NULL. Also fixed when make_envp was segfaulting
when there was an environment variable with NULL value (internal
variable).