Changed process creation and exit function
Fixed process creation to continue to following processes when
encountering an error in assigning a process. Changed
assign_process_error to essentially nullify a process when these errors
are encountered by setting argv to NULL and initialising in/outfile to
stdin/out. Changed multiple and single command functions to handle a
NULL argv. Also adjusted full exit to accept an int exit_status as a
parameter, rather than just taking the exit status from data structure -
to save lines of code. If full exit is required to take exit status from
the data structure it can be passed -1 as the exit status parameter.