Bug fixes
Bug fixes across project.
Fixed cd to update PWD if it is set, added in cd_utils.c
Fixed echo to correctly display '-' when it is given as the first
arugment.
Removed exit writing "exit" to stdout when executed, this is a setting
in other shells for debugging (set -v / set -x), to expand command and
print or to be verbose. Writing  "exit" is not a feature of the actual
exit command.
Fixed lots of issues with child processes not executing commands
correctly when PATH was not set, or if cmd was an empty string, or if
command was an executable in the current working directory. Should all
be fixed now.
Fixed segfaults with syntax errors of unclosed quotes or just empty
spaces input onto line.
Fixed tokeniser not skipping over tabspaces.
When expansion occurs into an empty string (char pointer to a single
null byte), execution should now occur correctly.