Child process fd leak fix
Fixed issue where child processes would leak fds when exiting early
(before execve), either due to a previous assigning process error
(invalid redirection, invalid cmd) or due to executing a built in. Fixed
by adding clean_child_fds function which closes all fds that are not
STDIN/STDOUT assigned to the current process which exited early, and all
remaining processes to be executed. This is only in the child process
which is exiting early, when the execution loop reaches following
processes their assigned fds will not be affected, as their copies in
the parent were unchanged.